refactor: reorganize subcommand into its own dir
