+++ src/doctor/checks/react_native.ts
