+++ src/doctor/core/projects.ts
