+++ src/doctor/utils/cache.ts
