+++ src/doctor/utils/helpers.ts
