+++ src/doctor/utils/patterns.ts
