+++ src/doctor/utils/string.ts
